chore: restructure project
