+++ src/api/routes/libp2p.ts
@@ -4,9 +4,9 @@ import { Ping } from '@libp2p/ping'
 import { multiaddr, Multiaddr } from '@multiformats/multiaddr'
 import { Request, Response, Router } from 'express'
 
+import { helia } from '../../services/helia/helia.js'
+import { logger } from '../../utils/logger.js'
 import { PeerIdDto } from '../dto/peer-id.dto.js'
-import { helia } from '../helia.js'
-import { logger } from '../utils/logger.js'
 
 const router = Router()
