+++ src/services/multer/multer.ts
@@ -1,8 +1,9 @@
 import multer from 'multer'
 
-import { config } from './config.js'
-import { ifs } from './helia.js'
-import { UnixfsMulterStorage } from './utils/unixfs-multer.storage.js'
+import { config } from '../../config.js'
+import { ifs } from '../helia/helia.js'
+
+import { UnixfsMulterStorage } from './unixfs-multer.storage.js'
 
 export const multerStorage = multer({
   storage: new UnixfsMulterStorage({
